use the server-side searchParams instead of client-side hook
